+++ test/builtin_test.exs
@@ -6,6 +6,7 @@ defmodule Interval.BuiltinTest do
   doctest Interval.Date, import: true
   doctest Interval.Float, import: true
   doctest Interval.DateTime, import: true
+  doctest Interval.Decimal, import: true
 
   test "Interval.Intervalable" do
     assert Interval.new(left: 1)
@@ -25,6 +26,24 @@ defmodule Interval.BuiltinTest do
       assert 4 === Interval.Integer.point_step(2, 2)
     end
 
+    test "size/1" do
+      assert 0 ===
+               Interval.Integer.new(left: 1, right: 1)
+               |> Interval.size()
+
+      assert 1 ===
+               Interval.Integer.new(left: 1, right: 2)
+               |> Interval.size()
+
+      assert nil ===
+               Interval.Integer.new(left: 1, right: nil)
+               |> Interval.size()
+
+      assert nil ===
+               Interval.Integer.new(left: nil, right: 2)
+               |> Interval.size()
+    end
+
     test "Interval.Intervalable" do
       assert Interval.Integer.new(left: 0, right: 1) ===
                Interval.new(left: 0, right: 1)
@@ -36,6 +55,24 @@ defmodule Interval.BuiltinTest do
       assert ~D[2022-01-03] === Interval.Date.point_step(~D[2022-01-01], 2)
     end
 
+    test "size/1" do
+      assert 0 ===
+               Interval.Date.new(left: ~D[2022-01-01], right: ~D[2022-01-01])
+               |> Interval.size()
+
+      assert 1 ===
+               Interval.Date.new(left: ~D[2022-01-01], right: ~D[2022-01-02])
+               |> Interval.size()
+
+      assert nil ===
+               Interval.Date.new(left: ~D[2022-01-01], right: nil)
+               |> Interval.size()
+
+      assert nil ===
+               Interval.Date.new(left: nil, right: ~D[2022-01-02])
+               |> Interval.size()
+    end
+
     test "Interval.Intervalable" do
       assert Interval.Date.new(left: ~D[2022-01-01], right: ~D[2022-01-02]) ===
                Interval.new(left: ~D[2022-01-01], right: ~D[2022-01-02])
@@ -45,6 +82,24 @@ defmodule Interval.BuiltinTest do
   describe "Float" do
     test "point_step/2" do
       assert nil === Interval.Float.point_step(2.0, 2)
+    end
+
+    test "size/1" do
+      assert 0.0 ===
+               Interval.Float.new(left: 1.0, right: 1.0)
+               |> Interval.size()
+
+      assert 1.0 ===
+               Interval.Float.new(left: 1.0, right: 2.0)
+               |> Interval.size()
+
+      assert nil ===
+               Interval.Float.new(left: 1.0, right: nil)
+               |> Interval.size()
+
+      assert nil ===
+               Interval.Float.new(left: nil, right: 2.0)
+               |> Interval.size()
     end
 
     test "Interval.Intervalable" do
@@ -64,6 +119,45 @@ defmodule Interval.BuiltinTest do
                right: ~U[2022-01-02 00:00:00Z]
              ) ===
                Interval.new(left: ~U[2022-01-01 00:00:00Z], right: ~U[2022-01-02 00:00:00Z])
+    end
+  end
+
+  describe "Decimal" do
+    test "point_step/2" do
+      assert nil === Interval.Decimal.point_step(Decimal.new(1), 2)
+    end
+
+    test "discrete?/1" do
+      refute Interval.Decimal.discrete?()
+    end
+
+    test "size/1" do
+      assert Decimal.new(0) ===
+               Interval.Decimal.new(left: Decimal.new(1), right: Decimal.new(1))
+               |> Interval.size()
+
+      assert Decimal.new(1) ===
+               Interval.Decimal.new(left: Decimal.new(1), right: Decimal.new(2))
+               |> Interval.size()
+
+      assert nil ===
+               Interval.Decimal.new(left: Decimal.new(1), right: nil)
+               |> Interval.size()
+
+      assert nil ===
+               Interval.Decimal.new(left: nil, right: Decimal.new(2))
+               |> Interval.size()
+    end
+
+    test "Interval.Intervalable" do
+      assert Interval.Decimal.new(
+               left: Decimal.new(1),
+               right: Decimal.new(2)
+             ) ===
+               Interval.new(
+                 left: Decimal.new(1),
+                 right: Decimal.new(2)
+               )
     end
   end
 end