+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Builtin `Jason.Encoder` support.
 - `Interval.__using__` option `jason_encoder` for including encoder. Defaults to `true`.
 - `Interval.to_map/1` to convert an Interval struct to a map suitible for JSON and similar serialization.
+- `Interval.Decimal` for `Decimal` support.
 
 ### Changed